refactor interactive task editing: extract `editTaskNoteInteractive` to internal package and improve task update logic
