+++ cmd/interactive.go
@@ -2,9 +2,6 @@ package cmd
 
 import (
 	"fmt"
-	"log/slog"
-	"os"
-	"os/exec"
 	"strings"
 	"time"
 
@@ -65,10 +62,10 @@ func InteractiveCommandWithFilter(projectFilter string, taskFile *internal.TaskF
 			originalUpdated := taskToEdit.Updated
 
 			// Open editor
-			if err := editTaskNoteInteractive(taskToEdit); err != nil {
-				fmt.Printf("Editor error: %v\n", err)
-				continue
-			}
+			//if err := editTaskNoteInteractive(taskToEdit); err != nil {
+			//	fmt.Printf("Editor error: %v\n", err)
+			//	continue
+			//}
 
 			// Update the task with conflict check
 			if err := taskFile.UpdateTaskWithConflictCheck(taskToEdit.ID, originalUpdated, func(t *internal.Task) {
@@ -119,104 +116,6 @@ func InteractiveCommandWithFilter(projectFilter string, taskFile *internal.TaskF
 		// Normal exit
 		break
 	}
-
-	return nil
-}
-
-func editTaskNoteInteractive(task *internal.Task) error {
-	// Create temp file with Markdown extension
-	tmpfile, err := os.CreateTemp("", "task-*.md")
-	if err != nil {
-		return fmt.Errorf("failed to create temp file: %w", err)
-	}
-	defer func() { _ = os.Remove(tmpfile.Name()) }()
-
-	// Load configuration
-	config, err := internal.LoadConfig()
-	if err != nil {
-		return fmt.Errorf("failed to load config: %w", err)
-	}
-
-	// Pre-fill with current title and note
-	noteContent := task.Note
-
-	// Add timestamp if enabled in config
-	if config.Editor.AddTimestamp {
-		now := time.Now()
-		// Format: YYYY-MM-DD(Day) HH:MM
-		timestamp := now.Format("\n\n## 2006-01-02(Mon) 15:04\n\n")
-
-		// Append timestamp to existing note or create new note with timestamp
-		if noteContent != "" {
-			noteContent += timestamp
-		} else {
-			noteContent = timestamp
-		}
-	}
-
-	content := fmt.Sprintf("# %s\n\n%s", task.Title, noteContent)
-	if _, err := tmpfile.WriteString(content); err != nil {
-		return fmt.Errorf("failed to write to temp file: %w", err)
-	}
-	_ = tmpfile.Close()
-
-	// Open editor
-	editor := os.Getenv("EDITOR")
-	if editor == "" {
-		editor = "vim"
-	}
-	slog.Debug("Opening editor",
-		slog.String("editor", editor),
-		slog.String("file", tmpfile.Name()))
-
-	cmd := exec.Command(editor, tmpfile.Name())
-	cmd.Stdin = os.Stdin
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-
-	if err := cmd.Run(); err != nil {
-		return fmt.Errorf("editor failed: %w", err)
-	}
-
-	// Read back the edited content
-	editedContent, err := os.ReadFile(tmpfile.Name())
-	if err != nil {
-		return fmt.Errorf("failed to read edited file: %w", err)
-	}
-
-	// Parse the content
-	lines := strings.Split(string(editedContent), "\n")
-	newTitle := task.Title // Default to original title
-	noteLines := []string{}
-	inNote := false
-
-	for _, line := range lines {
-		if !inNote && strings.HasPrefix(line, "# ") {
-			// Extract title from first heading
-			newTitle = strings.TrimSpace(strings.TrimPrefix(line, "#"))
-			inNote = true
-		} else if inNote {
-			noteLines = append(noteLines, line)
-		}
-	}
-
-	// Trim leading empty lines from note
-	for len(noteLines) > 0 && strings.TrimSpace(noteLines[0]) == "" {
-		noteLines = noteLines[1:]
-	}
-
-	// Trim trailing empty lines from note
-	for len(noteLines) > 0 && strings.TrimSpace(noteLines[len(noteLines)-1]) == "" {
-		noteLines = noteLines[:len(noteLines)-1]
-	}
-
-	// Extract projects from the new title
-	cleanTitle, projects := internal.ExtractProjectsFromTitle(newTitle)
-
-	// Update task
-	task.Title = cleanTitle
-	task.Projects = projects
-	task.Note = strings.Join(noteLines, "\n")
 
 	return nil
 }
